requirements.txt
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cors==1.0.1
cryptography==45.0.5
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
Flask-APScheduler==1.13.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jwt==1.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
PySocks==1.7.1
pytest==8.4.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
requests-file==2.1.0
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
tinycss2==1.4.0
tldextract==5.3.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
zope.event==5.1
zope.interface==7.2
